python-data-engineer/week8/exercises/tests/test_dms_operations.py
"""
DMS Operations Tests
---------------

Test cases for:
1. Resource management
2. Task operations
3. Monitoring
"""

import pytest
from unittest.mock import Mock, patch
import json
from datetime import datetime
from botocore.exceptions import ClientError

from ..dms_operations import DMSOperations

@pytest.fixture
def mock_dms_client():
    """Create mock DMS client."""
    with patch('boto3.client') as mock:
        client = Mock()
        mock.return_value = client
        yield client

@pytest.fixture
def dms_ops(mock_dms_client):
    """Create DMS operations instance."""
    return DMSOperations('us-west-2')

def test_create_replication_instance(dms_ops, mock_dms_client):
    """Test replication instance creation."""
    # Setup mock response
    mock_dms_client.create_replication_instance.return_value = {
        'ReplicationInstance': {
            'ReplicationInstanceIdentifier': 'test-instance',
            'ReplicationInstanceClass': 'dms.t3.micro',
            'ReplicationInstanceStatus': 'available'
        }
    }
    
    # Test instance creation
    instance = dms_ops.create_replication_instance(
        'test-instance',
        'dms.t3.micro',
        50
    )
    
    assert instance is not None
    assert instance['ReplicationInstanceIdentifier'] == 'test-instance'
    mock_dms_client.create_replication_instance.assert_called_once()
    
    # Test error handling
    mock_dms_client.create_replication_instance.side_effect = \
        ClientError({'Error': {'Code': '500', 'Message': 'Error'}}, 'operation')
    
    instance = dms_ops.create_replication_instance(
        'test-instance',
        'dms.t3.micro'
    )
    assert instance is None

def test_create_endpoints(dms_ops, mock_dms_client):
    """Test endpoint creation."""
    # Setup mock responses
    mock_dms_client.create_endpoint.side_effect = [
        {
            'Endpoint': {
                'EndpointIdentifier': 'source',
                'EndpointType': 'source',
                'EndpointArn': 'arn:source'
            }
        },
        {
            'Endpoint': {
                'EndpointIdentifier': 'target',
                'EndpointType': 'target',
                'EndpointArn': 'arn:target'
            }
        }
    ]
    
    # Test endpoint creation
    source_config = {
        'identifier': 'source',
        'engine': 'postgres',
        'settings': {
            'ServerName': 'source-db',
            'Port': 5432
        }
    }
    
    target_config = {
        'identifier': 'target',
        'engine': 'postgres',
        'settings': {
            'ServerName': 'target-db',
            'Port': 5432
        }
    }
    
    source_arn, target_arn = dms_ops.create_endpoints(
        source_config,
        target_config
    )
    
    assert source_arn == 'arn:source'
    assert target_arn == 'arn:target'
    assert mock_dms_client.create_endpoint.call_count == 2

def test_create_replication_task(dms_ops, mock_dms_client):
    """Test replication task creation."""
    # Setup mock response
    mock_dms_client.create_replication_task.return_value = {
        'ReplicationTask': {
            'ReplicationTaskIdentifier': 'test-task',
            'ReplicationTaskArn': 'arn:task'
        }
    }
    
    # Test task creation
    task_arn = dms_ops.create_replication_task(
        'test-task',
        'arn:instance',
        'arn:source',
        'arn:target',
        {'rules': []}
    )
    
    assert task_arn == 'arn:task'
    mock_dms_client.create_replication_task.assert_called_once()

def test_start_replication_task(dms_ops, mock_dms_client):
    """Test task start operation."""
    result = dms_ops.start_replication_task('arn:task')
    assert result == True
    
    # Test with start position
    result = dms_ops.start_replication_task(
        'arn:task',
        'checkpoint'
    )
    assert result == True
    
    # Verify calls
    assert mock_dms_client.start_replication_task.call_count == 2

def test_monitor_task(dms_ops, mock_dms_client):
    """Test task monitoring."""
    # Setup mock responses
    mock_dms_client.describe_replication_tasks.side_effect = [
        {
            'ReplicationTasks': [{
                'Status': 'running',
                'ReplicationTaskStats': {
                    'FullLoadProgressPercent': 50
                }
            }]
        },
        {
            'ReplicationTasks': [{
                'Status': 'stopped',
                'ReplicationTaskStats': {
                    'FullLoadProgressPercent': 100
                }
            }]
        }
    ]
    
    # Test monitoring
    dms_ops.monitor_task('arn:task', interval=1)
    assert mock_dms_client.describe_replication_tasks.call_count == 2

def test_cleanup_resources(dms_ops, mock_dms_client):
    """Test resource cleanup."""
    dms_ops.cleanup_resources(
        task_arn='arn:task',
        instance_arn='arn:instance',
        endpoint_arns=['arn:source', 'arn:target']
    )
    
    # Verify cleanup calls
    mock_dms_client.delete_replication_task.assert_called_once()
    assert mock_dms_client.delete_endpoint.call_count == 2
    mock_dms_client.delete_replication_instance.assert_called_once() 